package refactor, start to add some new auth/session classes
