+++ board-server/src/main/java/net/stzups/board/data/objects/TokenGenerator.java
@@ -1,0 +1,15 @@
+package net.stzups.board.data.objects;
+
+import java.security.SecureRandom;
+import java.util.Base64;
+
+public class TokenGenerator {
+
+    private static final SecureRandom secureRandom = new SecureRandom();
+
+    public static String generate(int length) {
+        byte[] bytes = new byte[length];
+        secureRandom.nextBytes(bytes);
+        return Base64.getEncoder().encodeToString(bytes);
+    }
+}
