+++ board-server/src/main/java/net/stzups/board/data/objects/User.java
@@ -8,21 +8,17 @@ import java.util.Random;
 public class User implements Serializable {
     private static Random random = new Random();
 
-    private HttpSession httpSession;
     private int id;
+    private byte[] hash;
     private List<String> ownedDocuments;
     private List<String> sharedDocuments;
 
-    public User(HttpSession httpSession) {
-        this.httpSession = httpSession;
+    public User(byte[] hash) {
         id = random.nextInt();
+        this.hash = hash;
         ownedDocuments = new ArrayList<>();
         //ownedDocuments.add(Document.createDocument(this).getId()); todo do somewhere else
         sharedDocuments = new ArrayList<>();
-    }
-
-    public HttpSession getHttpSession() {
-        return httpSession;
     }
 
     public int getId() {
