+++ board-server/src/main/java/net/stzups/board/data/objects/UserSession.java
@@ -1,0 +1,4 @@
+package net.stzups.board.data.objects;
+
+public class UserSession {
+}
