+++ board-server/src/main/java/net/stzups/board/server/ServerInitializer.java
@@ -1,8 +1,6 @@
 package net.stzups.board.server;
 
 import io.netty.channel.ChannelDuplexHandler;
-import io.netty.channel.ChannelHandler;
-import io.netty.channel.ChannelHandlerAdapter;
 import io.netty.channel.ChannelHandlerContext;
 import io.netty.channel.ChannelInitializer;
 import io.netty.channel.ChannelPipeline;
@@ -16,8 +14,9 @@ import io.netty.handler.stream.ChunkedWriteHandler;
 import io.netty.handler.traffic.GlobalTrafficShapingHandler;
 import io.netty.handler.traffic.TrafficCounter;
 import net.stzups.board.Board;
-import net.stzups.board.protocol.PacketEncoder;
-import net.stzups.board.protocol.PacketDecoder;
+import net.stzups.board.server.websocket.protocol.PacketEncoder;
+import net.stzups.board.server.websocket.protocol.PacketDecoder;
+import net.stzups.board.server.http.HttpServerHandler;
 
 import java.util.concurrent.Executors;
