+++ board-server/src/main/java/net/stzups/board/server/websocket/protocol/PacketDecoder.java
@@ -1,4 +1,4 @@
-package net.stzups.board.protocol;
+package net.stzups.board.server.websocket.protocol;
 
 import io.netty.buffer.ByteBuf;
 import io.netty.channel.ChannelHandler;
@@ -8,12 +8,12 @@ import io.netty.handler.codec.http.websocketx.BinaryWebSocketFrame;
 import io.netty.handler.codec.http.websocketx.TextWebSocketFrame;
 import io.netty.handler.codec.http.websocketx.WebSocketFrame;
 import net.stzups.board.data.objects.Point;
-import net.stzups.board.protocol.client.ClientPacket;
-import net.stzups.board.protocol.client.ClientPacketCreateDocument;
-import net.stzups.board.protocol.client.ClientPacketDraw;
-import net.stzups.board.protocol.client.ClientPacketHandshake;
-import net.stzups.board.protocol.client.ClientPacketOpenDocument;
-import net.stzups.board.protocol.client.ClientPacketType;
+import net.stzups.board.server.websocket.protocol.client.ClientPacket;
+import net.stzups.board.server.websocket.protocol.client.ClientPacketCreateDocument;
+import net.stzups.board.server.websocket.protocol.client.ClientPacketDraw;
+import net.stzups.board.server.websocket.protocol.client.ClientPacketHandshake;
+import net.stzups.board.server.websocket.protocol.client.ClientPacketOpenDocument;
+import net.stzups.board.server.websocket.protocol.client.ClientPacketType;
 
 import javax.naming.OperationNotSupportedException;
 import java.nio.charset.StandardCharsets;
