+++ board-server/src/main/java/net/stzups/board/server/websocket/protocol/server/ServerPacketRemoveUser.java
@@ -1,4 +1,4 @@
-package net.stzups.board.protocol.server;
+package net.stzups.board.server.websocket.protocol.server;
 
 import net.stzups.board.data.objects.User;
